t: O(s), s: O(s)
A palindrome is a string of characters in which `s == s.reversed`.

Given a string of characters, finding the longest possible palindrome is essentially finding the longest permutation of the characters in the string that also happens to be a palindrome. The problem with such an approach is that the number of char permutations w/o repetition of any string of length `s` is `N!`, and accumulating them in a container collection to check-for-palindrome adds a linear time component to each permutation, totalling `s * N!` per permutation of length `s`. Since we're searching for the longest `s`, we need to add to that expression, in the worst-case, another linear component to generate each length.

All put together, an upper bound to the total run time of such approach would be `O(s^2 * N!)` (not including the linear time to check-palindrome).

Another possible way of approaching the problem is thinking about the corollaries of `s == s.reversed`. A string equals its reverse iff you can fold the string from the middle and each side would have the same characters in a mirrored position. For this to hold, we either need characters that have an even number of appearances in total (half on each half of the string), or a single character in the middle position (forcing an odd length for the entire string). So, the longest possible palindrome of a set of characters is the result of taking all characters that appear an even number of times, plus all the odds minus one (because we need one character on each half of the string), plus one character in the middle.