t: O(n), s: O(k)
Deque solution. Linear time: at most, every element from 'nums' is added and removed only once from the queue.

Approach: Only keep track of the elements in the window that are required to return the maximum. Those elements are: [1] the maximum element in the first window (left index of the window is 0), and [2] any elements smaller than the maximum that appeared in the first window after the maximum.

So, we could use a deque keeping the invariants:
[1] all indices in the queue must be inside the current window (as the window moves right, drop the index that fall out of the window)
[2] for any element 'n' in the queue: any other element in the queue to its left is bigger or equal to 'n', and it appeared after (ie, to the right) in 'nums'

With this in mind, we simply need to update the queue holding this invariant as the window slides right. The left-most element in the queue will always be the current window maximum.

Let's look at an example to understand why this invariant solves the problem.

Assume this initial setup:
nums: [0 2 3] 1 0 1 2 3
queue: [3] <-- in the actual solution we store indices to drop them when they fall outside of the window as the window slides to the right.

The current maximum is 3. We don't need to store the 2 (or the 0) because the window moves right, so any future window that includes the 2 will also include the 3.

In the next window we have:
nums: 0 [2 3 1] 0 0 2 3
queue: [3 1]

In this case we need to keep track of the 1 because it may be the new maximum in a future window.

nums: 0 2 [3 1 0] 2 2 3
queue: [3 1 0]
nums: 0 2 3 [1 0 2] 2 3
queue: [2] <-- 3 is dropped first (out of window), then 1 and 0 and polled out because they are smaller than 2.