t: O(n*logn), s: O(n)
To reduce the task scheduler idle time we have to maximise the different tasks in a slot. One possible way of guaranteeing that condition is to fill in every interval (slot in the code) with the tasks with the highest remaining frequency without repetition in a single interval.

Simulation approach: heap + queue + map.

Time complexity: we're offering and polling each task from the heap exactly once => n tasks => n * log 26 <- we have, at most, 26 different tasks. A tighter upper bound would be n * log (#diff-tasks).