t: O(n), s: O(1)
Approach: To find out the total time the cpu will be running tasks, we can start by looking at the tasks that appear the most in the task array. If we call `m` the number of the most frequent tasks, we will require at least `m-1` full intervals to complete them. It's `m-1` because the last interval doesn't require the cpu to idle.
`n` is the number of idle cycles the cpu will have to idle to run same task, so we can consider `n+1` to be the length of an interval where no duplicate task can appear.
The total time required up to this point is `(n+1) * (m-1)`. For the remaining tasks (those with frequency < maxFreq), we need to calculate how many of them we can fit in the empty slots of the intervals left by those with maxFreq and add those that didn't fit. This is equivalent to comparing the length of the tasks array with the time required to complete all of them.