t: O(n), s: O(1)
A string `s` is a palindrome <=> `rev(s)` == `s`. In this case we cannot reverse `s` because: [1] the challenge requires O(1) space, and [2] reversing a list in-place changes the original list by definition, so we wouldn't be able to iterate over both to see if the condition is satisfied.

Instead, we can reverse half of the list in-place, split both halves, and compare them. The solution implements this approach.