t: O(1), s: O(1)
The number of operations required is the accumulated difference between each element in the first (or second) half of the array and the element in the middle of the array. More specifically, [1] if the length of the array is even, it's the accumulated difference between each element and the average of the two middle elements; [2] if it's odd, it's the acc. diff. with the middle element.
[1] follows the arithmetic sequence: 1 3 5 7 ...
[2] follows the arithmetic sequence: 2 4 6 8 ...

https://www.mathsisfun.com/algebra/sequences-sums-arithmetic.html