t: O(n), s: O(n)
This one is a little tricky to explain, but I'll try my best. In `sum` I keep track of the prefix summation (ie, the cumulative value of all the elements from 0 to `i`). So, intuitively, `if sum == k` means that the elements in the subarray `[0, i]` will sum k, and it should be counted.

Now, here's the tricky part. In `subarrSums` I store all the values of `sum` to search all the prefix sums up to `i` in O(1), and consider the number of occurrences.
I search this hash for `sum-k`. If `sum-k` is in the map, it means that there's at least one subarray whose elements sum up to k. Let's look at an example for clarification.

[3 4 6 1 ... ] arr
[3 7 13 14] prefixsum
sum -> 3 7 13 14
k -> 7

When `i == 1`, `sum == 7` and `sum == k`, which means that the subarray [0, 1] sums k by our definition of `sum`.
When `i == 3`, `sum-k == 7` and `prefixsum contains 7`. The only way this can happen is if you have a subarray whose elements sum up to k. In order to understand this, you need to remember that with prefixsum you're keeping track of the deltas of adding each element `i` to the subarray `[0, i]`. So, finding `sum-k` in prefixsum is equivalent to finding a subarray `[j, i], j > 0` whose elements have added `k` to the current global summation.

Credit: https://www.youtube.com/watch?v=HbbYPQc-Oo4