t: O(n*h), s: O(h^2)
h: height of the tree => if the three is balanced => h = log_2 (n)
Time complexity: Visit every node. On each node: visit all open paths until that node, which is the depth of the node
Space complexity: h recursive calls. Each rec. call holds a reference to the list of open paths of size h